Add files via upload
These are the necessary files for TASK 3 of the Week 13 completion. These files create a new animated health power-up with scripts that make it spawn in a random location--after a short delay it will disappear and destroy itself on its own--and once the player grabs it, it will destroy the health object and add 1 life to the life counter WITHOUT exceeding 3 lives. As a bonus, there is also a "gain life" sound effect that plays once the player and health object collide.