Add files via upload
These files are dedicating to animating, creating, and scripting a coin in the plane game. The coin works to appear on screen for a couple of seconds and then disappear, and once the player collides with the coin it will add 1 to the score and then destroy itself. It also plays a coin sound.